create store for cph search page and include CPH warnings create seperate component for autocomplete Input box to avoid rerendering whole component while searching
